# Optimize cart list in customer detail (a bit)