--- a/src/Adapter/Customer/QueryHandler/GetCustomerForViewingHandler.php
+++ b/src/Adapter/Customer/QueryHandler/GetCustomerForViewingHandler.php
@@ -302,28 +302,24 @@ final class GetCustomerForViewingHandler implements GetCustomerForViewingHandler
      */
     private function getCustomerCarts(Customer $customer)
     {
-        $carts = Cart::getCustomerCarts($customer->id);
+        $result = Db::getInstance(_PS_USE_SQL_SLAVE_)->executeS('
+        SELECT c.id_cart, c.date_add, ca.name as carrier_name, c.id_currency, cu.iso_code as currency_iso_code
+        FROM ' . _DB_PREFIX_ . 'cart c
+        LEFT JOIN ' . _DB_PREFIX_ . 'carrier ca ON ca.id_carrier = c.id_carrier
+        LEFT JOIN ' . _DB_PREFIX_ . 'currency cu ON cu.id_currency = c.id_currency
+        WHERE c.`id_customer` = ' . (int) $customer->id . '
+        ORDER BY c.`date_add` DESC');
+        
         $customerCarts = [];
-
-        foreach ($carts as $cart) {
-            $cart = new Cart((int) $cart['id_cart']);
-            Context::getContext()->cart = $cart;
-
-            $currency = new Currency($cart->id_currency);
-            Context::getContext()->currency = $currency;
-
-            $carrier = new Carrier($cart->id_carrier);
-            $summary = $cart->getSummaryDetails();
-
+        foreach ($result as $row) {
+            $cart = new Cart((int) row['id_cart']);
             $customerCarts[] = new CartInformation(
-                sprintf('%06d', $cart->id),
-                Tools::displayDate($cart->date_add, null, true),
-                $this->locale->formatPrice($summary['total_price'], $currency->iso_code),
-                $carrier->name
+                sprintf('%06d', $row['id_cart'])
+                Tools::displayDate($row['date_add'], null, true),
+                $this->locale->formatPrice($cart->getOrderTotal(true), $row['iso_code']),
+                $row['carrier_name']
             );
         }
-
-        Context::getContext()->currency = Currency::getDefaultCurrency();
 
         return $customerCarts;
     }