| Questions         | Answers
| ----------------- | -------------------------------------------------------
| Branch?           | develop
| Description?      | I speed up the cart listing on customer page a little bit. More info below.
| Type?             | improvement
| Category?         | BO
| BC breaks?        | no
| Deprecations?     | no
| Fixed ticket?     | Fixes #25683 - only partial improvement, do not close
| How to test?      | Create many carts, and then open "View customer" page. Compare loading time before and after.
| Possible impacts? | Open customer detail and see that it works the same. Enable debug profilling in config/defines.ing and see that query count is reduced.

### Investigation
When customer has few orders, the page is slow AF. It's only slow because of carts listing.
Test scenario: 58 orders, 60 carts, 92 bought products, FAST SERVER

### Numbers of queries and time
| Case         | Queries | Time
| ----------------- | ----------------- | -----------------
| Original | 2070 | 2 seconds
| With my PR | 1720 | 1,4 seconds
| Without cart listing at all, but everything else | 73 | 0,3 seconds
| With cart listing at all, but no cart total | 74 | 0,3 seconds
| With cart listing, creating cart object, price formatting, but replacing `$cart->getOrderTotal(true)` for hardcoded number | 135 | 0,35 seconds

### Solutions
- Completely refactor cart calculation
- Cache cart total after every calculation to ps_cart table
- Remove cart total from all cart listings


<!-- Click the form's "Preview" button to make sure the table is functional in GitHub. Thank you! -->

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/prestashop/prestashop/26689)
<!-- Reviewable:end -->