@Hlavtox 2021-11-22
Thanks @matthieu-rolland! Guys please, can you try to think about the way to optimize this?
`getOrderTotal` is a total performance killer.