@marwachelly 2021-12-15
Hello, 
I created 142 orders and carts in 2 shops (1 shop with PR and 1 shop without PR) : 
the following behaviour is observed 
**with PR :** 

![image](https://user-images.githubusercontent.com/89604350/146195313-60a8975a-3428-408f-a1ea-637197c707b2.png)

**without PR:**
![image](https://user-images.githubusercontent.com/89604350/146196146-fb7e784e-5ec8-472e-a800-63faf76be9cf.png)

Could you please check?

Thanks,